refactor: atoms の移動
