+++ src/components/atoms/ConfigMenu/MenuStack/MenuStack.tsx
@@ -1,7 +1,7 @@
 import clsx from 'clsx';
 import { useCallback, useMemo } from 'react';
 import { useRecoilState } from 'recoil';
-import { atomOfAnswerImageVariant } from '../../../atoms';
+import { atomOfAnswerImageVariant } from '../../../../atoms';
 import Button from '../../Button';
 import './MenuStack.scss';
